// udhd-mis-front-end-main/src/app/audit/internal-audit/models/revCapExpDTO.ts
export class revCapExpDTO {

  fy1L!: string;
  fy2L!: string;
  fy3L!: string;
  iaId!: number;
  id!: any;
    administrativeExpenses!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;

    };
    details!: {
      fy1: string;
      fy2: string;
      fy3: string;
      fy4: string;
      fy5: string;
      fy6: string;
      iaId: number;
      id: any;
    };
    developmentWorks!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;
    };
    loanRepayment!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;
    };
    operationAndMaintenance!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;
    };
    otherCapitalExpenditure!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;
    };
    otherExpenditure!: {
      expenditureType: string;
      fy1Amt: number;
      fy2Amt: number;
      fy3Amt: number;
      fy4Amt: number;
      fy5Amt: number;
      fy6Amt: number;
     id: any;
      // type: string;
    }

}